\name{AgingPred}
\alias{}
\title{Aging Prediction}
\usage{
AgingPred()
}
\description{
Execute "AgingPred()", then a software window will pop up. In the window, upload the corresponding file to obtain the predicted age for each sample and the proportion of aging rate for each group.
}
\examples{
AgingPred()
}
